+++ ctp/src/ctp_market_spi.rs
@@ -65,7 +65,6 @@ impl Rust_CThostFtdcMdSpi_Trait for Spi {
 }
 
 fn _convert_tick(market_data: &CThostFtdcDepthMarketDataField) -> Tick {
-
     let tick = Tick {
         symbol: c_char_to_string(market_data.InstrumentID.as_ptr()),
         datetime: format!("{} {}", c_char_to_string(market_data.ActionDay.as_ptr()), c_char_to_string(market_data.UpdateTime.as_ptr())),
@@ -74,7 +73,7 @@ fn _convert_tick(market_data: &CThostFtdcDepthMarketDataField) -> Tick {
         high: market_data.HighestPrice,
         low: market_data.LowestPrice,
         close: market_data.ClosePrice,
-        volume: market_data.Volume,
+        volume: market_data.Volume as f64,
         turnover: market_data.Turnover,
         open_interest: market_data.OpenInterest,
         last_price: market_data.LastPrice,
@@ -83,21 +82,21 @@ fn _convert_tick(market_data: &CThostFtdcDepthMarketDataField) -> Tick {
         bid_price3: market_data.BidPrice3,
         bid_price4: market_data.BidPrice4,
         bid_price5: market_data.BidPrice5,
-        bid_volume1: market_data.BidVolume1,
-        bid_volume2: market_data.BidVolume2,
-        bid_volume3: market_data.BidVolume3,
-        bid_volume4: market_data.BidVolume4,
-        bid_volume5: market_data.BidVolume5,
+        bid_volume1: market_data.BidVolume1 as f64,
+        bid_volume2: market_data.BidVolume2 as f64,
+        bid_volume3: market_data.BidVolume3 as f64,
+        bid_volume4: market_data.BidVolume4 as f64,
+        bid_volume5: market_data.BidVolume5 as f64,
         ask_price1: market_data.AskPrice1,
         ask_price2: market_data.AskPrice2,
         ask_price3: market_data.AskPrice3,
         ask_price4: market_data.AskPrice4,
         ask_price5: market_data.AskPrice5,
-        ask_volume1: market_data.AskVolume1,
-        ask_volume2: market_data.AskVolume2,
-        ask_volume3: market_data.AskVolume3,
-        ask_volume4: market_data.AskVolume4,
-        ask_volume5: market_data.AskVolume5,
+        ask_volume1: market_data.AskVolume1 as f64,
+        ask_volume2: market_data.AskVolume2 as f64,
+        ask_volume3: market_data.AskVolume3 as f64,
+        ask_volume4: market_data.AskVolume4 as f64,
+        ask_volume5: market_data.AskVolume5 as f64,
     };
     tick
 }
